get working suspense boundary, add skeleton
